fix delete recette bug

+++ src/Controller/CourseController.php
@@ -166,8 +166,12 @@ class CourseController extends AbstractController
     #[Route('/{id}', name: 'app_course_delete', methods: ['POST'])]
     public function delete(Request $request, Course $course, CourseRepository $courseRepository): Response
     {
-        if ($this->isCsrfTokenValid('delete'.$course->getId(), $request->request->get('_token'))) {
-            $courseRepository->remove($course, true);
+        $connectedUser = $this->security->getUser();
+
+        if($recette->getUser() == $connectedUser){
+            if ($this->isCsrfTokenValid('delete'.$course->getId(), $request->request->get('_token'))) {
+                $courseRepository->remove($course, true);
+            }
         }
 
         return $this->redirectToRoute('app_course_index', [], Response::HTTP_SEE_OTHER);
